updated player, tiger, tree, ...

+++ Scorpion.h
@@ -8,7 +8,7 @@ private:
     
 public:
     Scorpion();
-    Scorpion(int x, int y);
+    Scorpion(float x, float y);
     ~Scorpion();
 
 };
